[canaria] HomeComponent で管理していた auth ロジックを LoginComponent に委譲
