+++ projects/canaria/src/app/pages/home/home.component.html
@@ -1,8 +1,1 @@
-<div *ngIf="user$ | async as user; else showLogin">
-  <h1>Hello {{ user.displayName }}!</h1>
-  <button (click)="logout()">Logout</button>
-</div>
-<ng-template #showLogin>
-  <p>Please login.</p>
-  <button (click)="login()">Login with Google</button>
-</ng-template>
+<p>home works!</p>
