+++ projects/canaria/src/app/pages/home/home.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { Authenticator } from 'utilities';
 
 @Component({
   selector: 'app-home',
@@ -7,16 +6,6 @@ import { Authenticator } from 'utilities';
   styleUrls: ['./home.component.scss'],
 })
 export class HomeComponent implements OnInit {
-  constructor(private authenticator: Authenticator) {}
-  readonly appTitle = 'Canaria';
-  user$ = this.authenticator.loggedInUser$;
-
-  login() {
-    this.authenticator.login();
-  }
-  logout() {
-    this.authenticator.logout();
-  }
-
+  constructor() {}
   ngOnInit() {}
 }
