+++ projects/canaria/src/app/pages/login/login.component.html
@@ -1,1 +1,8 @@
-<p>login works!</p>
+<div *ngIf="user$ | async as user; else showLogin">
+  <h2>Hello {{ user.displayName }}!</h2>
+  <button (click)="logout()">Logout</button>
+</div>
+<ng-template #showLogin>
+  <p>Please login.</p>
+  <button (click)="login()">Login with Google</button>
+</ng-template>
